not for land: one way to provide BC for ConvPackedParams
Summary:

This is a hack to allow ConvPackedParams to load data from
previous versions. The implementation adds a check for
a custom `convertLegacyFormat` function (name is debug, will need
finalizing).  If the function exists, it passes the input through
that function before passing it to `__getstate__`.

Code is not clean and ready for landing, ideally we can align
on whether this approach is valid before finalizing the code,
and merge this into the preceding PR if needed.

Test Plan:

Hack test script to verify from user's POV:
https://gist.github.com/vkuzo/f3616c5de1b3109cb2a1f504feed69be

The script does the following:
* saves a quantized conv (needs to be run on a revision without this stack)
* loads a quantized conv (needs to be run on this revision)

before this PR the script fails, after this PR it succeeds.

real tests would be added if this is the approach people want to take

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 234393c2114e5f47d2727afc19a18e3de615d35e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43087